Update class.widget.js
Update javascript to support nulling out connected widgets when the table data changes and previously selected items/hosts are not in the new table data. This is very important because if a graph is connected to this widget, but you use the host navigator to change the host group or host of this widget, the graph does not update. This can be confusing to users